rename map component file
